Added Makefile and .gitignore
Created a simple Makefile to simplify the building process. This
makefile is expected to be modifed or replaced as development continues.

With this makefile the project can be built by executing `make` and the
project directory can be cleaned by executing `make clean`.

Additionally, added a .gitignore file so all files created with make are
automatically ignored from `git status`.